… californium-proxy/pom.xml
… californium-proxy/src/main/java/org/eclipse/californium/proxy/HttpTranslator.java
… californium-proxy/src/main/java/org/eclipse/californium/proxy/resources/ForwardingResource.java
… californium-proxy/src/main/java/org/eclipse/californium/proxy/resources/ProxyCacheResource.java
… californium-proxy/src/main/java/org/eclipse/californium/proxy/resources/ProxyCoapClientResource.java
… californium-proxy/src/main/java/org/eclipse/californium/proxy/resources/ProxyHttpClientResource.java