@martinmine 2017-11-17
Before this gets merged: Both me and @vikram919 have had some issues regarding the dependencies. Im building the Californium framework by running `mvn clean install`, however by doing so the produced JARs does not contain the dependencies that is required to use the framework, and these does then have to be added explicitly together with whatever project is then referencing the JARs. Is this intentional? I have a strong feeling I am missing something here.