@vikram919 2017-11-17
@martinmine 
Thanks for your information, after adding the dependency it worked for now, I will let you know If I know the solution for above question.
One question as you described the test setup:

 `HTTP requestster -- <HTTP> -- Californium -- <CoAP> -- Californium -- <HTTP> -- nginx`

 as from the example we can acheieve 

`HTTP requester --- <HTTP> --Californium Server`

how  can we establsih?

`Coap requester ----<CoAP>----HTTP Server`