@boaks 2017-11-17
Thanks!
But including GPL dependency is, as fa as I know, not possible for a eclipse project.
(I hope, I haven't overseen this in your first PR.)

For the "httpasyncclient" we (the committers) must apply a CQ at eclipse for that.
I'm out of the office until 27.11. but will work on your PR, when I'm back.