@martinmine 2017-11-17
@boaks Thanks, sorry, I was not aware of this regarding the licenses.
On another note, I just re-visited the benchmarks on the 2.x branch.
Against the commit before the changes (850f6f3d1fc33f73405685e114200acebb64d14f): Avg: 3.47ms, stdDev: 1.09ms, N: 86399
Against the changes in this PR: Avg: 1.42ms, stdDev: 1ms, N: 210529