@boaks 2017-11-17
@sophokles73 
Maybe, you can already ask eclipse about the "httpasyncclient" (CQ)