@martinmine 2017-11-17
@vikram919 I wrote my setup based on the proxy example in the repo. Requests are then made towards the coap2http resource. There is also one small change I made in CoapTranslator.java to make it work (for some reason it wasn't setting the proxy URIs on the requests:
```diff
diff --git a/californium-proxy/src/main/java/org/eclipse/californium/proxy/CoapTranslator.java b/californium-proxy/src/main/java/org/eclipse/californium/proxy/CoapTranslator.java
index 6ef83f6e..4d81c92a 100755
--- a/californium-proxy/src/main/java/org/eclipse/californium/proxy/CoapTranslator.java
+++ b/californium-proxy/src/main/java/org/eclipse/californium/proxy/CoapTranslator.java
@@ -90,6 +90,7 @@ public final class CoapTranslator {
 
 		// get the uri address from the proxy-uri option
 		URI serverUri;
+		String proxyUriForward = null;
 		try {
 			/*
 			 * The new draft (14) only allows one proxy-uri option. Thus, this
@@ -97,6 +98,13 @@ public final class CoapTranslator {
 			 */
 			String proxyUriString = URLDecoder.decode(
 					incomingRequest.getOptions().getProxyUri(), "UTF-8");
+
+			final String proxyEndpoint = "coap2http/";
+			final int proxyEndpointPos = proxyUriString.indexOf(proxyEndpoint);
+			if (proxyEndpointPos > 0) {
+				proxyUriForward = proxyUriString.substring(proxyEndpointPos + proxyEndpoint.length());
+				proxyUriString = proxyUriString.substring(0, proxyEndpointPos + proxyEndpoint.length());
+			}
 			serverUri = new URI(proxyUriString);
 		} catch (UnsupportedEncodingException e) {
 			LOGGER.warn("UTF-8 do not support this encoding: {}", e);
@@ -121,6 +129,11 @@ public final class CoapTranslator {
 		options.removeBlock2();
 		options.clearUriPath();
 		options.clearUriQuery();
+
+		if (proxyUriForward != null && proxyUriForward.length() > 0) {
+			options.setProxyUri(proxyUriForward);
+		}
+
 		outgoingRequest.setOptions(options);
 		
 		// set the proxy-uri as the outgoing uri
```