@martinmine 2017-11-17
@boaks I've rebased and re-implemented the CompletableFuture class under the compat package. Should be good from the GPL license now :)