@vikram919 2017-11-20
Hi @martinmine 
Thanks for the help with the code.
I have created a test case:
`Coap Client` -->` Californium<coap2coap>` -->`Californium<coap2http>` --> `HTTP local Server`
 and calculated the average after exchanging requests 1000 times it is 8.432ms
I have not tried with the same test case as yours.
I will put it here as soon as I have them.