@martinmine 2017-11-20
@vikram919 Do you also have some warmup period/request in your test? What was the standard deviation in your result? In my setup, I run as many requests as possible through the system for 10 minutes on one single thread requesting a HTML document and one warmup request.