@martinmine 2017-11-20
Btw, what system are you testing it on? (CPU/OS/memory)