@vikram919 2017-11-20
> Do you also have some warmup period/request in your test? What was the standard deviation in your result?

SD is 3.105ms, I see at the beginning it started with 16ms (initially).

> Btw, what system are you testing it on? (CPU/OS/memory)

CPU: Intel 4600M, i7 2.90GHz processor
16GB RAM Windows OS