@vikram919 2017-11-20
@martinmine 
As @sophokles73 mentioned,
This dependency is alone okay, it works now without including it again in cf-proxy pom.xml.
we can remove the rest.

```

<dependency>
	<groupId>org.apache.httpcomponents</groupId>
	<artifactId>httpasyncclient</artifactId>
	<version>4.1.3</version>
	<scope>compile</scope>
</dependency>

```