@martinmine 2017-11-20
@vikram919 @sophokles73 Thanks for your comments, I will address them shortly.