@martinmine 2017-11-20
@sophokles73 Maybe you should add the thing about the copyright headers to the contributors.md file as well? :) Regarding the exceptions, I highly appreciate (as a developer when using Californium) that I get a more appropriate message out rather than just for example "Null pointer exception". A stack trace is soooo much more useful if no other exception is thrown. For future work, I would suggest that you go by the convention of providing the Throwable object to the logger so it can print the stack trace (if any).