@vikram919 2017-11-21
@martinmine 
I created a test setup without changing the code just only with the pull request,
`Coap client`<--`<coap2http>`-->`HTTP local Server` - Average 0,236708 milli seconds for a each request-response
`Http client`<--`<HttpProxy>`-->`Coap local Server` - Average 0,865336 milli seconds

This time I run the tests using a single thread sending 1000 get requets and calculating delay on every response  in approx nano time precision with the same machine mentioned above
Thank you for the good work!