@martinmine 2017-11-21
@vikram919 Thanks for your investigation, did you run it without the changes from this PR as well?