@martinmine 2017-11-21
@vikram919 Sorry, if I was unclear. Did you run your benchmark on the head of the branch `eclipse:2.0.x` so that we know how much of a performance boost this PR gives? According to your results, we only know how it performs with the changes proposed in this PR.