@sophokles73 2017-11-22
@martinmine can you please squash your changes into one or two commits, e.g. one adapting the logging and the other implementing the improvements?