@martinmine 2017-11-22
@sophokles73 Done. I squashed it into one commit as I had commits between log-related changes that I didn't want to start mess with. Hope this is fine.