@sophokles73 2017-11-22
@martinmine 
thanks for squashing, but you should fix the commit message which now contains all messages of the squashed commits including the "signed-off-by" footer. You should at least remove the duplicate footers so that there is only one single "signed-off-by" at the bottom. Maybe you could also arrange the changes in a bullet point list? Or include some excerpts from the excellent description you created for the PR?