@martinmine 2017-11-22
@sophokles73 Thanks. Do you think it looks better now?