@vikram919 2017-11-24
@martinmine 
Did you try using observe option over coap2coap proxy?