@martinmine 2017-11-27
@vikram919 are you talking about this? https://github.com/martinmine/californium/blob/8c0fbe3cf92d99a3e6e3385471fdd662b41edea7/californium-proxy/src/main/java/org/eclipse/californium/proxy/resources/ProxyCoapClientResource.java#L78 Otherwise, I use the proxy example in the current master.